requirements.txt
termcolor
nats-py
boto3
python-dotenv
instaloader
pytube